requirements.txt
fastapi>=0.100.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
backports.zoneinfo>=0.2.1;python_version<"3.9"
